requirements.txt
streamlit
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
pandas
openpyxl
fpdf2
plotly
kaleido
